+++ ChangeLog.md
@@ -1,3 +1,8 @@
+## 0.1.1.0
+
+* Adds support for setuid and setguid when running command
+* Adds support for setting current directory when running command
+
 ## 0.1.0.1
 
 * Turn off all RTS options